pyproject.toml
[tool.poetry]
name = "lista-django"
version = "0.1.0"
description = ""
authors = ["Plinio Schmidt de Arruda Junior <pliniojunior5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.0"
python-magic = "^0.4.27"
Pillow = "^9.2.0"
drf-spectacular = "^0.24.2"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
